+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix bug preventing request.is_preview from being set
 - Hide wagtail userbar on preview
+- Make sure in_preview_panel is set
 
 ### Removed