+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug preventing request.is_preview from being set
 - Hide wagtail userbar when in preview panel
 - Make sure in_preview_panel is set
+- Drop non running get_preview_url override from base page (@marteinn)
 
 ### Removed