+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hide wagtail userbar when in preview panel (@marteinn)
 - Make sure in_preview_panel is set (@marteinn)
 - Drop non running get_preview_url override from base page (@marteinn)
+- Remove deprecated wagtail.contrib.modeladmin from INSTALLED_APPS (@marteinn)
 
 ### Removed